Implement the faster shadow stuff from : http://git.iodoom.org/~raynorpat/iodoom3/raynorpats-glsl_iodoom3/commit/332fa0d46f68e4057a00275a75e09c1f3b1ab7c8#comment_24
Realistically I dont think it pushes out performance up by much, however it might on lower end hardware. One thing to note however is that the game feels a lot smoother with this change, however it might be down to the new glext header - as a good few things have changed. In any case - test this out and let me know.

git-svn-id: https://svn.thedarkmod.com/svn/darkmod_src/trunk@5237 49c82d7f-2e2a-0410-a16f-ae8f201b507f