4520: Make sure scripting works in x64 bit mode. Also, now savegames are compatible between 32-bit and 64-bit builds.
In case when pointer was stored in script storage, now it is encoded and stored as 32-bit number.
In 32-bit mode, encoding of a pointer is simply its address.
In 64-bit mode, encoding of a pointer is its offset relative to base address of the special memory zone.
Since all such pointers point into idScriptObject::data, we allocate all these memory chunks with special memory allocator.
The allocator itself is implemented in Heap_Embedded.*.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@6879 49c82d7f-2e2a-0410-a16f-ae8f201b507f