Fixed bug (most likely) with sound propagation initialization. Detected with TypeInfo Memory Debugging =)
So the core problem is that the check like this:
  if ( (area->portals = new SsndPortal[np]) == NULL )
may hold true even on successful allocation, when np is zero.
And this happens (at least on Inn Business).
As a result, error handling code triggers, and most of m_sndAreas remain uninitialized.

The fix is: removed all the erroneous checks for success of memory allocation.
There is simply no reason to do them; if we exhaust memory, no actions could save us from crash anyway.
Additionally, check pointers for null before copying and init portalDists to null (most likely these fixes are not necessary).


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@7024 49c82d7f-2e2a-0410-a16f-ae8f201b507f