Added utility functions for old intermediate opengl crap like qglColor, use these if you feel the need for further checks. They are clamped to reasonable values as default.
Also added wrappers for scissor and viewport,
which can be modified if the need arises.

Also left some commented console prints in a function that was modified heavily and which i reverted to original, 
these also tell why it should not be touched.
Uncomment "slightly humerous" console prints as you see fit to see for yourself.
Just remember to turn them off again, or else the engine will slow to a crawl.

Revelator signing off.

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@7638 49c82d7f-2e2a-0410-a16f-ae8f201b507f