Console history small change
When entering a new command and it exists in the history the old occurrences are now removed. 
This way it won't allow duplicates any more.

History is now an idStrList instead of the fixed array of idEditField. 
Why? Easier access to the string elements by index. Built-in AddUnique() function. One less index member and less maintenance code.

The in-game console history is now unlimited in elements count. When written out to disk, it's limited to 100 strings instead of 64 (no reason but makes it more 'human')

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@7743 49c82d7f-2e2a-0410-a16f-ae8f201b507f