#5028. Post EV_UpdateCameraTarget event only for entities having camera target.
Note: I have also bumped maximum number of events from 8K to 10K.
This is done in order to avoid such nasty situations, when number of entities fits the limit, but there are a bit more events, and loading stops working.
The 8K should swallow the "one-event-per-entity" spam if it is present, and the added 2K should sustain a bit of added stuff.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@8191 49c82d7f-2e2a-0410-a16f-ae8f201b507f