#5311. Assign name as initial value of variable only for string variables.
There is no point in assigning the name to a numeric variable, it won't be assigned anyway, since it is a string (not starting with a digit).
Note that string variables are idWinStr, and idWinBackground inherited from it.
All the rest are numeric (integer, float, or float vector).

This is necessary in order to add warnings on bad assignments in the following commit.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@8903 49c82d7f-2e2a-0410-a16f-ae8f201b507f