GLFW: use custom recipe, don't disable fPIC.
The stock recipe did not pass CONAN settings to CMAKE (via conan_basic_setup).
As the result, it could not be used to cross-compile 32-bit library on 64-bit Linux, which was pretty inconvenient.

Disabling fPIC via default options is bad because it causes error on Windows.
Remember that almost every recipe simply deletes fPIC option on Windows.
I think it is OK to compile libs as PIC, especially if they are not performance-critical.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@8965 49c82d7f-2e2a-0410-a16f-ae8f201b507f