#5138. Extracted code to compute number of particles in particle deform system.
The new function is:
  int idParticle_GetParticleCountOnSurface(const idPartStageData &stg, const srfTriangles_t *tri, float totalArea, int &particleCountPerCycle);

Note that it returns two integers.
In case of "particle deform" (uniform distrubution on surface), the numbers are equal.
In case of "particle deform2" (uniform distribution by triangles):
  particleCountPerCycle is number of particles per one triangle
  return value is the total number of particles in the system (particleCountPerCycle * number of triangles)

The particleCountPerCycle must be passed to subsequent idParticle_* functions and is required for correct bunching computations and for fadeIndex feature.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@9016 49c82d7f-2e2a-0410-a16f-ae8f201b507f