#5439. idBounds::FromPointRotation/FromBoundsRotation: fixed wrong results. Especially when for rotation angle > 180 degrees.
Note that this problem was very easy to detect (before this fix).
Grab some object, and rotate it around player.
When sweeping over some specific angle, you see a noticeable lag.
That's because at this problematic frame, the bounds are wrong and large, so collision with the whole world is checked.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@9157 49c82d7f-2e2a-0410-a16f-ae8f201b507f