#5528. TDM no longer crashes at map load when a Response has been set on the security camera.
idSecurityCamera::spawn contained the following, which occurs in many other spawnclasses too:

// SR CONTENTS_RESPONSE FIX
if( m_StimResponseColl->HasResponse() )
	physicsObj.SetContents( physicsObj.GetContents() | CONTENTS_RESPONSE );

physicsObj seems not to be initialised, so according to debug mode, TDM crashes at GetContents() when a security camera with a Response is spawned.
Replacing physicsObj. with GetPhysics()-> fixes the crash and allows the security camera to Respond to stims.

A similar problem might exist with other spawnclasses, too, since the code block above is quite common.

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@9176 49c82d7f-2e2a-0410-a16f-ae8f201b507f