#5562. Optimize concatenation of ambient triangle groups in WriteOutputSurfaces in dmap.
Due to something like OOP stupidity, the code called MergeTriLists repeatedly to concatenate a small list to the back of the growing result list.
The function traverses the destination list to find its end (yeah, it is singly-linked, so end pointers is not stored).
With N lists of length K each, this takes O(N^2 K) time.

The new code saves pointer to the last node between concatenations, reducing time to O(N K).
There is no cvar to disable this optimization, since it surely does not change behavior.

Hot dmap time improvements:
  painterswife\city: ???
  NHAT\politics: ?1:05? -> 0:44


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@9204 49c82d7f-2e2a-0410-a16f-ae8f201b507f