Added meta-cvar dmap_compatibility for compiling old maps.
I have a habit of putting dmap_XXX cvar for everything which might affect output (as much as I can).
This is true both for the serious changes of 2.08 and for optimizations of 2.10.
The problem is that there are too many individual cvars to mess with.

The idea is that if someone sets this cvar to desired TDM version,
than dmap resets all of its individual cvars in such way that behavior is closest to that version.
This is obviously much easier to use for mappers and developers.

The main use case is setting it to 207, since it reverts major changes in visportal/opacity systems.

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@9216 49c82d7f-2e2a-0410-a16f-ae8f201b507f