#5118. Removed artificial five-second delay after the FM is loaded.
The idea is that some FMs display special tip when the mission if loaded --- probably remains of times when progress bar stayed at 100% for a lot of time.
The problem is that this slows down loading, and does not work anyway, since the pause applies to black screen.
Also, it breaks mission start if you did not succeed at the first time due to error.

More details here:
  https://bugs.thedarkmod.com/view.php?id=5118
  https://forums.thedarkmod.com/index.php?/topic/16332-the-mission-loading-process/&do=findComment&comment=458233


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@9231 49c82d7f-2e2a-0410-a16f-ae8f201b507f