#4550. Removed hacky CPUID code from Simd.cpp.
This code was added long time ago when Doom 3 was closed source, in order to workaround lack of CPUID detection on Linux.
Now CPUID detection code works same was both on Linux and Windows, so this additional detection code it unnecessary.

With this commit, newer SIMD implementations like AVX and AVX2 should work properly on Linux!

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@9253 49c82d7f-2e2a-0410-a16f-ae8f201b507f