Improving FPS counter to avoid frequent switching between close values.
Note that this has little effect with default value of com_showFpsAvg (N = 4).
But when it is large (e.g. N = 100), then it is still typical for average value to change every frame.

Now FPS counter is updated to the current value of one of the following happens:
1) N/2 frames have passed. So with N = 100 you have guaranteed update every 50 frames.
2) Average FPS has changed by more than 3% (and by more than 1) since last update.
If none of these conditions are satisfied, then old average FPS is displayed.



git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@9266 49c82d7f-2e2a-0410-a16f-ae8f201b507f