#4550. Do not compile SSE/AVX SIMD processors on non-x86 architectures.
Added macro ENABLE_SSE_PROCESSORS, which is enabled if target arch is x86 or SSE2 is enabled in compiler (some compilers may cross-compile SSE intrinsics).
All the methods of SSE/AVX processors are compiled under this macro flag, and creation of SSE/AVX processors is also undeer this flag.

This is needed in order to compiler TDM on CPUs like ARM or PowerPC.
Note that such platforms are not supported by TDM team.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@9278 49c82d7f-2e2a-0410-a16f-ae8f201b507f