#5354 & #5592. Preliminary refactoring for upcoming dmap improvements.
1) Record original brush which generated every interAreaPortal_s.
This would be helpful in info_locationseparator diagnostics, so that we can print brush number if visportal has any issues.

2) Extracted FindLeafNodeAtPoint from PlaceOccupant.
Finding BSP leaf by point is a commonly used function, needed to get the node that entity belongs to.

3) FindShortestPathThroughBspNodes extended in many ways:
  a) Can start from many nodes at once now. It would be needed for new leak path detection (#5592).
  b) Can return set of all visited nodes. This would be needed to mark all reachable areas in a separate array (#5354).
  c) Callbacks now have context. The "CanPass" functions are now lambdas, and LambdaToFuncPtr is used to pass them as function pointers.

4) FindPortalCycleBFS renamed to ReportDroppedPortal.
Since now it not only finds the problem, but also reports it in all possible ways.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@9285 49c82d7f-2e2a-0410-a16f-ae8f201b507f