#5592. Replaced the algorithm which searches for leak path (report LESS points).
The old algorithm uses Depth-First Search and some ad-hoc magic in order to generate leak path.
This algorithm does not guarantee getting shortest path.
In fact, it usually obtains a very long path going through the whole level.

The new algorithm uses Breadth-First Search called "FindShortestPathThroughBspNodes" --- the same as used for visportals and locations diagnostics.
It surely finds the shortest path --- the one having the least number of points (i.e the least number of BSP portal transitions).
Note that in this case we start BFS from all the nodes containing entities at once: all of them are put into initial queue with zero distance.

The code for reporting leaks moved from leakfile.cpp to FloodEntities function in portals.cpp.
That's where all the other code which saves pointfiles is located.
As a result, function ProcessModel in dmap.cpp is simplified, and leakfile.cpp is removed.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@9288 49c82d7f-2e2a-0410-a16f-ae8f201b507f