#5591. Do not use TexStorage + TexSubImage for compressed textures of weird sizes.
OpenGL spec requires TexSubImage updates of compressed textures to not cross 4x4 blocks.
The case when we upload full texture of size not divisible by 4 seems to violate this requirement (padding is inevitably done by compressor).
While it works on NVIDIA, it crashes on AMD GPU.

Since this is merely a loading time optimization, let's just use a slightly slower but safer TexImage in such cases.

See also:
  https://forums.thedarkmod.com/index.php?/topic/20980-crash-when-loading-irregular-sized-images-with-amd-gpu/
  https://forums.thedarkmod.com/index.php?/topic/21073-a-house-of-locked-secrets-crashes-tdm-on-startup/&tab=comments#comment-464686
  

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@9579 49c82d7f-2e2a-0410-a16f-ae8f201b507f