#5815. Changed r_spotlightBehavior cases, description, and default behavior.
First of all, now the behavior of Doom 3 (and TDM up to 2.07) is the default one.

Second, there are three modes of behavior like in Doom 3 BFG, with minor variations of how far clipping plane of the frustum are defined.
Suppose that N and F are distance to "light_start" and "light_end" points along target direction.
Then we have different position of the far clipping plane: at F, at F + N, and at F + N^2/(F+N).


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@9666 49c82d7f-2e2a-0410-a16f-ae8f201b507f