#5815. Extracted ShadowAtlasForVector into tdm_shadowmaps.glsl. Also lowered version on volumetric to GLSL 3.3.
Note that now we have yet another define TDM_allow_ARB_texture_gather.
Unfortunately it is required in order to comditionally declare a function in a common include, because:
  1) #extension directive cannot be inside include, since it must go before anything else in the shader.
  2) If you don't enable the #extension, then but textureGather gives compile error, but GL_ARB_texture_gather is 1 for some reason =(
  3) It seems wrong to force every shader including tdm_shadowmaps.glsl to enable texture gather #extension regardless of whether they need it or not.

To be honest, it would be much easier to simple remove texture gather usage from everywhere.
It does not make the jaggies any less visible anyway.



git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@9681 49c82d7f-2e2a-0410-a16f-ae8f201b507f