Do not blit-resolve stencil FBO if there were no shadows to render.
This greatly improves performance with both soft stencil shadows and multisampling.
It seems that quite a lot of lights behave that way, and copying whole FBO so many times eats a lot of time.

BTW, one possible further optimization is to blit only the part of stencil FBO according to light scissor.
This would improve performance with multisampling when there are many small lights on screen.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@9698 49c82d7f-2e2a-0410-a16f-ae8f201b507f