Yet another attempt at fixing language selection, fixing the mess from 9837.
Svn rev 9837 caused all the menu sounds to disappear after engine restart:
  https://forums.thedarkmod.com/index.php?/topic/21197-devs-210-is-branched-out/&do=findComment&comment=469712

It happened because "reloadLanguage" command tried ot restart engine when sound system is not fully initialized yet (no soundCache).
As seen in idSoundShader::ParseShader, sound samples are silently thrown away if soundCache is not available, which means they will never be loaded in the future when the cache is created.
I think this if was done in order to support "s_noSound 1" case, where soundCache is never created at all.
I added assert that soundCache exists in idSoundShader::ParseShader to avoid any missing sounds in future.
BTW, the problematic loads of sounds shaders happen inside loading of main menu GUI.

The reason that happened is that I18NLocal::SetLanguage could not distinguish between cases: "full engine reload" and "user changed language in menu".
If user changes language, then we should destroy and recreate main menu GUI, but if whole engine is reloaded, that we should not do that (main menu GUI will be recreated later).
So the main fix is in idSessionLocal::Shutdown: main menu GUI is now destroyed when session is shutdown.
There is no reason to keep menu alive when session is dead...
As the result, now I18NLocal::SetLanguage sees that main menu GUI does not exist and it does not try to mess with GUI (during engine reload).

Finally, I added the assert that mission manager exists in idSessionLocal::CreateMainMenu.
I handled the NULL case in rev 9837, but now it is clear that main menu should not be created so early, and this should never happen.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@9838 49c82d7f-2e2a-0410-a16f-ae8f201b507f