#5648. Don't process each edge twice in idBrush::AddBevelsForAxialBox (under dmap_aasExpandBrushUseEdgesOnce cvar).
Note that brush is represented as a set of its faces, each face described sequence of points.
The points are not necessarily exactly equal, this we need some numeric magic to detect that edges are logically the same.

When we consider an edge on face A, we determine the other face B as the one giving maximum signed distance (with face A excluded from consideration).
This criterion should be pretty stable.

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@9898 49c82d7f-2e2a-0410-a16f-ae8f201b507f