gameLocal.activeEntities is now an array (with NULLs) instead of linked list.
This slightly improves performance of game modelling on large maps.
For instance of the first beta of Lucy's Quest, the change is:
  RunGameTic: {11.17, 10.99, 10.9} -> {10.46, 10.57, 10.69}     (not much but measurable)
  DeactivateEntities: 0.24 -> 0.035                             (huge improvement)
  SortActiveEntityList: 0.56 -> 0.31                            (seems to become much better)

game/EntityList.cpp
game/EntityList.h
  That's data structure for storing list of entities.
  Supports adding to the end, removing entities, and iterating through whole list.
  Note: relies on int member of idEntity storing a back index in it.

1) idGameLocal::activeEntities are now of type idEntityList.
Of course, idEntity now stores "int activeIdx" instead of "idLinkList activeNode".
Also, fixed the code for iterating over activeEntities everywhere.

2) Removed boolean flags sortPushers and sortTeamMasters from idGameLocal.
On sufficiently large maps both flags are true most of the time, so there is no point in this rarely triggered "optimization".

3) idGameLocal::SortActiveEntityList fully reimplemented.
Instead of doing 3 passes with "move to front" on special entities,
now this function classifies all entities into 6 buckets based on same properties.
Then the buckets are concatenated to produce a new list of active entities.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@9900 49c82d7f-2e2a-0410-a16f-ae8f201b507f