Allocate large arrays in idProgram on heap, instead of being a part of global variable.
It is now important that capacity of variables, functions, statements stay equal to MAX_GLOBALS, MAX_FUNCS, MAX_STATEMENTS respectively.
I added a bunch of asserts to make sure that this is true.

The reason for doing this is that idProgram used to be more than 4 MB in size, and it is aggregated into idGameLocal, which exists as a global variable gameLocal.
As the result, having at least one member of idGameLocal with nonzero initial value increases size of Linux executable by 4 MB:
  https://forums.thedarkmod.com/index.php?/topic/21377-fun-with-member-pointers/&do=findComment&comment=473045


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@9906 49c82d7f-2e2a-0410-a16f-ae8f201b507f