#5954. Optimized spatial data structure in idClip.
The old data structure was fixed BSP tree, which splits the world bounds into 16 x 16 x 16 cells,
with linked list of clipmodels attached to every cell (BSP leaf).
Non-adaptive space subdivision and linked lists caused ClipModelsTouching(Moving)Bounds run slow due to cache/TLB misses.

This new data structure idBoxOctree stores dynamic object/boxes of varying sizes.
It is an octree, with every node (not only leaves) containing a list of objects in it.
The tree is grown on demand: when a node gets more than 90 objects which are smaller than 25% of node box size, 8 sons are generated and small objects are pushed down to them.
This allows the structure to naturally adapt to the shape and density of the particular map.
The attached list of objects is stored as a linked list of 127-object chunks, which combines fast traversal of arrays and simple custom memory allocation.

On Lucy's Quest first beta, I see the following improvement in timings:
  Fresh start (1 min):
    9.62 -> 9.14  RunGameTic
    1.16 -> 1.04  ProcessStimResponse
    0.66 -> 0.63  ServiceEvents
  Savegame load (5 min):
    7.82 -> 7.52  RunGameTic
    1.09 -> 0.96  ProcessStimResponse
    0.64 -> 0.62  ServiceEvents
This map has a lot of AIs, each of them regularly sends visualStim to the others.
In order to send visualStim, idClip query is done on a pretty large box.
This stresses ClipModelsTouchingBounds function inside ProcessStimResponse.


idlib/containers/FlexList.h
  Added Pop method like in idList.
idlib/bv/Bounds.h
  Added Expand/ExpandSelf by vector.

idlib/bv/BoxOctree.*
  This is a new data structure used in idClip.

game/physics/Clip.*
  Now idClip uses the new idBoxOctree data structure to store clipmodels.
  idClipModel::Link now uses slightly faster "Update" method for changing position, instead of Remove + Add.
  World bounds are expanded to cubic size before passing to octree (previous code used them as is, which resulted in stretched cells).
  Absolute bounds of every clipmodel are stored near the pointer to it, so ClipModelsTouching(Moving)Bounds can prune objects geometrically without touching memory of idClipModel.

game/Game_local.cpp
  Added calls to idClip::Optimize = idBoxOctree::Condense when game is loaded (both new game and loading a save).
  Unfortunately, many clipmodels are initially inserted into idClip at (0,0,0) instead of their correct position.
  The idClip::Optimize partly reverts the harm done by this practice (although not completely).


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@9918 49c82d7f-2e2a-0410-a16f-ae8f201b507f