#5076. Extracted pk4 patterns match into static method of Pk4Mappings and added fast implementation for trivial regexes.
We have 7000 individual files, and each of them is checked against hundreds of regexes, most of which are a/b/c or ^a/b/c.
Pk4Mappings::SearchOne checks if regex is a trivial one and does plain substring search or prefix match for them.

Note that MSVC's implementation of std::regex is very slow, especially in Debug.
It is a torture to wait for SortFilesIntoPk4s phase to finish if you want to debug something that happens later.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@9964 49c82d7f-2e2a-0410-a16f-ae8f201b507f