#5828 !!! Specular term of interaction shader no longer depends on diffuse texture contents.
Previously specular term had a weird modulation by (75% + 25% * diffuseTexColor).
That's of course in addition to modulation by specular color parameter.

It is not worth hacking the standard Phong model with such things.
If artist wants this effect, he can blend diffuse texture into specular texture (even using image programs).

This commit changes the visual look of the game, and thus is a likely candidate for revert in future.

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10089 49c82d7f-2e2a-0410-a16f-ae8f201b507f