#6083. Implemented algorithm to find a point inside area, and teleportArea command.
The new algorithm is idRenderWorld::GetPointInArea, which is intended only for developer usage.
It considers two bounding boxes:
  1) of the "_areaN" model      (sometimes empty)
  2) same as 1, but also with all adjacent visportals added to box    (sometimes much larger than the area)

For each box, it builds 13 x 13 x 13 grid of points.
For each point, it finds distance to closest obstacle in 6 axis-aligned directions, and computes a "score" out of these distances.
The distance is found using new function idRenderWorldLocal::RecurseFullLineIntersectionBSP_r, which raytraces BSP tree and finds all intervals of intersection.
The point with best score wins.

If there is no winner for box 1, then box 2 is considered.
If no winner is found at all, then middle of box 1 or box 2 is returned (not surely inside area).
If all boxes are empty, then we report error and return nothing.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10098 49c82d7f-2e2a-0410-a16f-ae8f201b507f