Added idBitArray container and utility functions.
This is similar to bitset from STL/boost.
Except that we support static functions accepting raw pointers, which means that you can manage lifetime as you see fit and apply bit operations on top of that.

The following routines are included:
1) single bit: get, set, invert
2) segment: set to same value, invert
3) insert given bitmask to an interval (as long as interval is shorter than bitset word)
We can add more in future if needed (e.g. IfBitsAnyTrue + IfBitsAnyFalse, CountBits).

I also added a stress test, which should pretty reliably verify that this code works correctly.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10173 49c82d7f-2e2a-0410-a16f-ae8f201b507f