Don't cache glScissor, just reset it every single time.
The problem is that scissor is cached in backend.currentScissor not inside some wrapper (like GL_State or GL_SelectTexture), but manually at every call site.
Needless to say, not all call sites do this checking, so this is very error-prone.
At the same time, calling glScissor is most likely cheap, so we don't need to bother.

This was done to understand why screenshots had wrong scissor with r_softShadowsMipmaps:
  https://forums.thedarkmod.com/index.php?/topic/21662-timing-of-211-beta/&do=findComment&comment=480815
It did not fix the problem, but is still a good idea =)

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10184 49c82d7f-2e2a-0410-a16f-ae8f201b507f