Take savegame screenshot at full current resolution, then downscale it to suitable size.
FBO management in TDM is pretty complicated, and doing custom-resolution rendering is hard.
This damn "crop" functionality should better be removed completely, since it requires support at every part of rendering code =(
Instead of doing low-resolution screenshot, it is much more reliable to take standard-resolution screenshot and then postprocess it.

So we take default-reslution screenshot.
Since it is in RGB, we convert it to RGBA.
Then downscale it using R_MipMap until it becomes less than 1/4 of FullHD resolution.
This screenshot has good quality (because it usually has 4x supersampling) and very small size.

This fixes the issue with tiled optimization of soft stencil shadows.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10197 49c82d7f-2e2a-0410-a16f-ae8f201b507f