#5766. Fixed idFileSystemLocal::AddUnique to be case-insensitive. Note that the comparison is already case-insensitive here, but hash function is not... so this was probably the original intent.
This function is called from GetFileList in order to collect a list of files by filter.
Since idTech4 virtual filesystem is case-insensitive (even on Linux), we should always report file at most once.
But having case-sensitive comparison in AddUnique could result in file reported several times with different casing.

One particular example when this happens is when someone unpacks PK4 archive with uppercase filename inside: packed and unpacked files will have different case.
That's because VFS normalizes all filenames inside PK4 to lowercase.
For unpacked files, it reports file as it is.
Note that when file is searched, all comparisons are case-insensitive: even on Linux the engine searches for a file which matches.
(except for directories which differ by case --- that won't work on Linux)

Right now this fixes a bug that if we fully unpack cos3_sacricide, then we get error "Decl file '%s' is loaded twice" on skins/Bienie_models_architecture_modules2.skin.


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10217 49c82d7f-2e2a-0410-a16f-ae8f201b507f