#6231. Fixed delayed player damage when a fall ends with a mantle (by Daft Mugi).
Previously player velocity was not changed when mantle started, until the very end of the mantle.
That was quite strange, because player obviously stopped and started going through fixed mantling animation, while still having huge negative Z velocity...
Now it is set to zero immediately when mantle is started.

Note that as side product, the damage in this case is increased.
If the fall velocity is -V, then damage was f(D - 150), while now it is proper f(D).
That's because before zeroing velocity, it was first set to -150 due to some totally unrelated hacks.

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10263 49c82d7f-2e2a-0410-a16f-ae8f201b507f