#6271. Smoothen transition from "very fogged" to "totally fogged" in fog shader.
To be honest, I could not notice the difference.
But I still believe fogging depend on everything smoothly, without jumps.

Also note that this adds difference against the original Doom 3 fog behavior!
Doom 3 has a jump from factor = 0.991 to factor = 1.
But as I said, the difference is hardly visible anyway.

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10369 49c82d7f-2e2a-0410-a16f-ae8f201b507f