#6271. Optimized soft particles fragment shader for high-overdraw cases.
Basically, pass inverse values in uniforms, and replace divisions with multiplications.
The only two divisions that remain are inside depthToZ and they can't be optimized.

Note that I don't like such low-level optimizations, but particles are often rendered with insane overdraw.
And this microoptimization helps to restore the old performance in "Hidden Hands: The Lost Citadel".

See also:
  https://forums.thedarkmod.com/index.php?/topic/21796-6271-backends-merge/&do=findComment&comment=484928

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10372 49c82d7f-2e2a-0410-a16f-ae8f201b507f