#6271. Fixed bug with volumetric light shadows in single-pass shadow maps mode.
The volumetric lights are now force-switched to shadow maps even when stencil shadows are normally used.
The single-pass mode is used when both r_shadowMapSinglePass = 1 and r_shadows = 2, i.e. it does NOT trigger in case of such volumetric lights.
However, the per-light shadow maps code was checking only for r_shadowMapSinglePass = 1 for a skip, so it did not work either.

This is fixed, and now there is only one place where singlepass flag is checked.

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10393 49c82d7f-2e2a-0410-a16f-ae8f201b507f