dmap ear-cutting: consider zero-angle vertices as convex.
This idea comes form FIST paper about robust polygon triangulation with ear clipping.
Obviously, angles with cross product = 0 can have angle of 0, 180, and 360 degrees.
The case of 180 degrees is easily determined by dot product sign.
The cases of 0 and 360 degrees are distinguished by going along the vertex chain and looking to which side it first exists the line.

The case of 0 angle can be cut too.
The case of 360 angle... well, we should not cut it, although cutting it does not seem to make any harm either =)


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10397 49c82d7f-2e2a-0410-a16f-ae8f201b507f