#6300. Deleted unused image formats from functions used in image reporting.
Removed texture formats 1,2,3,4.
These ones are supported by old OpenGL 1.0, but never used in our code.

Removed LATC compressed format.
It was probably added as an experiment, but now DXT + RGTC are the only compressed formats used.

Removed GL_COMPRESSED_RGB(A)_ARB format.
This is "generic" compressed format, i.e. OpenGL driver is free to decide which compression it wants to use.
We don't use this, and I think we should always control space and error of compression.

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10402 49c82d7f-2e2a-0410-a16f-ae8f201b507f