#6300. idMath::FtoiFast now uses round-to-nearest as originally indended.
More specifically, it uses current FPU rounding mode, which should always be round-to-nearest with ties resolved to even.
There are three implementations:
 * fast SSE2-only
 * fast C++ emulation (fast math compiler setting would break it)
 * standard C++ function which is very slow on MSVC
All of them provide exactly the same result on normal floats in the range of int.

Originally, this function contained custom implementations with round-to-nearest behavior (based on x87 fistp).
And for some reason it also contained C++ fallback with round-to-zero behavior =(
Then in svn rev 6842 (year 2017) I disabled custom code for x64 compatibility and thus switched behavior from round-to-nearest to round-to-zero.
The fact that this function has the same name as Ftoi (which always rounds to zero) did not help to notice the change.

So this commit changes behavior of FtoiFast back from round-to-zero to round-to-nearest.
This function is actively used throughout the code, so it might cause some issues.

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10410 49c82d7f-2e2a-0410-a16f-ae8f201b507f