#6300. Refactored SIMD initialization infrastructure.
I want to compare generic implementation against SSE2 one in image compression unit tests.
To achieve this, I need to easily and silently replace SIMD processor temporarily.
The current architecture does not support that.

Now there is idSIMD::CreateProcessor function, which generates processor silently.
It can give processor with specified name or just the default one.
This function can be used to create a temporary processor, then std::swap it with "SIMDProcessor" global variable.

The weirdness with "processor" and "generic" global variables is removed.
And the code is simpler and more logical in general.

git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10419 49c82d7f-2e2a-0410-a16f-ae8f201b507f