Make console reports in "r_showportals 2" mode more clear/useful.
It no longer prints data for lightgem and subviews, only one line is printed for player view.
That's what most people want I guess, and multiline output every frame is very hard to read.

Also, now the area where camera/eye is located is printed first.
Previously it could be guessed as the first number in any of the space-separated groups =)

Finally, now '->' and ',' separators are used instead of '-' in every group.
So it becomes clear that the first number in group is special (area we consider), and the other numbers comprise a list (areas connected to it by portals).

And tweaked coloring a bit =)


git-svn-id: https://svn.thedarkmod.com/publicsvn/darkmod_src/trunk@10427 49c82d7f-2e2a-0410-a16f-ae8f201b507f